+++ motech-server-core/src/main/java/org/motechproject/server/ws/RegistrarWebService.java
@@ -77,7 +77,7 @@ public class RegistrarWebService implements RegistrarService {
 			@WebParam(name = "iptDose") Integer iptDose,
 			@WebParam(name = "iptReactive") Boolean iptReactive,
 			@WebParam(name = "itnUse") Boolean itnUse,
-			@WebParam(name = "fht") Integer fht,
+			@WebParam(name = "fht") Double fht,
 			@WebParam(name = "fhr") Integer fhr,
 			@WebParam(name = "urineTestProteinPositive") Boolean urineTestProteinPositive,
 			@WebParam(name = "urineTestGlucosePositive") Boolean urineTestGlucosePositive,
@@ -282,8 +282,8 @@ public class RegistrarWebService implements RegistrarService {
 			@WebParam(name = "ttDose") Integer ttDose,
 			@WebParam(name = "lochiaColour") Integer lochiaColour,
 			@WebParam(name = "lochiaAmountExcess") Boolean lochiaAmountExcess,
-			@WebParam(name = "temperature") Integer temperature,
-			@WebParam(name = "fht") Integer fht,
+			@WebParam(name = "temperature") Double temperature,
+			@WebParam(name = "fht") Double fht,
 			@WebParam(name = "comments") String comments)
 			throws ValidationException {
 
@@ -364,7 +364,7 @@ public class RegistrarWebService implements RegistrarService {
 			@WebParam(name = "referred") Boolean referred,
 			@WebParam(name = "maleInvolved") Boolean maleInvolved,
 			@WebParam(name = "weight") Double weight,
-			@WebParam(name = "temperature") Integer temperature,
+			@WebParam(name = "temperature") Double temperature,
 			@WebParam(name = "bcg") Boolean bcg,
 			@WebParam(name = "opv0") Boolean opv0,
 			@WebParam(name = "respiration") Integer respiration,
@@ -410,8 +410,8 @@ public class RegistrarWebService implements RegistrarService {
 			@WebParam(name = "vitaminA") Boolean vitaminA,
 			@WebParam(name = "dewormer") Boolean dewormer,
 			@WebParam(name = "weight") Double weight,
-			@WebParam(name = "muac") Integer muac,
-			@WebParam(name = "height") Integer height,
+			@WebParam(name = "muac") Double muac,
+			@WebParam(name = "height") Double height,
 			@WebParam(name = "maleInvolved") Boolean maleInvolved,
 			@WebParam(name = "comments") String comments)
 			throws ValidationException {
@@ -577,7 +577,7 @@ public class RegistrarWebService implements RegistrarService {
 			@WebParam(name = "motechId") Integer motechId,
 			@WebParam(name = "ancRegNumber") String ancRegNumber,
 			@WebParam(name = "estDeliveryDate") Date estDeliveryDate,
-			@WebParam(name = "height") Integer height,
+			@WebParam(name = "height") Double height,
 			@WebParam(name = "gravida") Integer gravida,
 			@WebParam(name = "parity") Integer parity,
 			@WebParam(name = "enroll") Boolean enroll,
